+++ robot-core/src/main/java/org/obolibrary/robot/ExportOperation.java
@@ -1,7 +1,6 @@
 package org.obolibrary.robot;
 
 import com.google.common.collect.Lists;
-import java.io.*;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -942,11 +941,12 @@ public class ExportOperation {
       ShortFormProvider provider = col.getShortFormProvider();
 
       Cell cell;
-      switch (colName) {
+      switch (colName.toUpperCase()) {
         case "IRI":
           String iriString = entity.getIRI().toString();
-          if (format.equalsIgnoreCase("html")) {
-            String display = String.format("<a href=\"%s'\">%s</a>", iriString, iriString);
+          if (format.toLowerCase().startsWith("html")) {
+            String display =
+                String.format("<a href=\"%s\">%s</a>", iriString, iriString.replace("&", "&amp;"));
             cell = new Cell(col, display, iriString);
           } else {
             cell = new Cell(col, iriString);
@@ -968,7 +968,7 @@ public class ExportOperation {
           List<String> values = getSynonyms(ontology, entity);
           row.add(new Cell(col, values));
           continue;
-        case "subclasses":
+        case "SUBCLASSES":
           if (entity.isOWLClass()) {
             Collection<OWLClassExpression> subclasses =
                 EntitySearcher.getSubClasses(entity.asOWLClass(), ontology);