+++ robot-core/src/main/java/org/obolibrary/robot/ManchesterOWLSyntaxObjectHTMLRenderer.java
@@ -61,7 +61,7 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            property.getIRI().toString(), getShortFormProvider().getShortForm(property)));
+            property.getIRI().toString(), escape(getShortFormProvider().getShortForm(property))));
   }
 
   /** Given an OWLClass, write a hyperlink describing it to the writer. */
@@ -70,7 +70,7 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            cls.getIRI().toString(), getShortFormProvider().getShortForm(cls)));
+            cls.getIRI().toString(), escape(getShortFormProvider().getShortForm(cls))));
   }
 
   /** Given an OWLDataProperty, write a hyperlink describing it to the writer. */
@@ -79,7 +79,7 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            property.getIRI().toString(), getShortFormProvider().getShortForm(property)));
+            property.getIRI().toString(), escape(getShortFormProvider().getShortForm(property))));
   }
 
   /** Given an OWLDataType, write a hyperlink describing it to the writer. */
@@ -88,7 +88,7 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            node.getIRI().toString(), getShortFormProvider().getShortForm(node)));
+            node.getIRI().toString(), escape(getShortFormProvider().getShortForm(node))));
   }
 
   /** Given an OWLNamedIndividual, write a hyperlink describing it to the writer. */
@@ -97,7 +97,8 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            individual.getIRI().toString(), getShortFormProvider().getShortForm(individual)));
+            individual.getIRI().toString(),
+            escape(getShortFormProvider().getShortForm(individual))));
   }
 
   /** Given an OWLObjectProperty, write a hyperlink describing it to the writer. */
@@ -106,7 +107,7 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
     write(
         String.format(
             "<a href=\"%s\">%s</a>",
-            property.getIRI().toString(), getShortFormProvider().getShortForm(property)));
+            property.getIRI().toString(), escape(getShortFormProvider().getShortForm(property))));
   }
 
   /**
@@ -159,5 +160,15 @@ public class ManchesterOWLSyntaxObjectHTMLRenderer extends ManchesterOWLSyntaxOb
           ce.toString(),
           ce.getClass().toString());
     }
+  }
+
+  /**
+   * Get an HTML-safe rendering of a string by replacing & and <
+   *
+   * @param str String to escape
+   * @return HTML-safe string
+   */
+  private String escape(String str) {
+    return str.replace("&", "&amp;").replace("<", "&lt;");
   }
 }